# Conflicts:
#	src/main/java/gregtech/GT_Mod.java
#	src/main/java/gregtech/common/GT_Client.java
#	src/main/java/gregtech/loaders/postload/recipes/Pulverizer.java

+++ src/main/java/gregtech/api/metatileentity/implementations/GT_MetaTileEntity_Transformer.java
@@ -175,10 +175,11 @@ public class GT_MetaTileEntity_Transformer extends GT_MetaTileEntity_TieredMachi
             for (byte i = 0; i < 6 && aBaseMetaTileEntity.getStoredEU() < aBaseMetaTileEntity.getEUCapacity(); i++)
                 if (aBaseMetaTileEntity.inputEnergyFrom(i)) {
                     TileEntity tTileEntity = aBaseMetaTileEntity.getTileEntityAtSide(i);
-                    if (tTileEntity instanceof IEnergyProvider && ((IEnergyProvider) tTileEntity)
-                            .extractEnergy(ForgeDirection.getOrientation(GT_Utility.getOppositeSide(i)), 1, true)
-                            == 1) {
-                        long tEU = (long) ((IEnergyProvider) tTileEntity).extractEnergy(
+                    if (tTileEntity instanceof IEnergyProvider && ((IEnergyProvider) tTileEntity).extractEnergy(
+                            ForgeDirection.getOrientation(GT_Utility.getOppositeSide(i)),
+                            1,
+                            true) == 1) {
+                        long tEU = ((IEnergyProvider) tTileEntity).extractEnergy(
                                 ForgeDirection.getOrientation(GT_Utility.getOppositeSide(i)),
                                 GT_Utility.safeInt(maxEUInput() * 100L / GregTech_API.mRFtoEU),
                                 false);
@@ -186,7 +187,7 @@ public class GT_MetaTileEntity_Transformer extends GT_MetaTileEntity_TieredMachi
                         aBaseMetaTileEntity.injectEnergyUnits((byte) 6, Math.min(tEU, maxEUInput()), 1);
                     } else if (tTileEntity instanceof IEnergyStorage
                             && ((IEnergyStorage) tTileEntity).extractEnergy(1, true) == 1) {
-                                long tEU = (long) ((IEnergyStorage) tTileEntity).extractEnergy(
+                                long tEU = ((IEnergyStorage) tTileEntity).extractEnergy(
                                         GT_Utility.safeInt(maxEUInput() * 100L / GregTech_API.mRFtoEU),
                                         false);
                                 tEU = tEU * GregTech_API.mRFtoEU / 100;
@@ -271,7 +272,8 @@ public class GT_MetaTileEntity_Transformer extends GT_MetaTileEntity_TieredMachi
             IWailaConfigHandler config) {
         final int facing = getBaseMetaTileEntity().getFrontFacing();
         final NBTTagCompound tag = accessor.getNBTData();
-        final int side = (byte) accessor.getSide().ordinal();
+        final int side = (byte) accessor.getSide()
+                                        .ordinal();
         final boolean allowedToWork = tag.getBoolean("isAllowedToWork");
 
         final byte inputTier = GT_Utility.getTier(tag.getLong("maxEUInput"));
